+++ src/components/index.ts
@@ -1,5 +1,3 @@
 export { Header } from './Header';
-export { Button } from './Button';
-export { Select } from './Select';
 export { TextEditor } from './TextEditor';
 export { ImageUploadPreview } from './ImageUploadPreview';
